Restructured for better style and layout switching
Styles and layouts can now be swapped easily with a URL parameter.
Styles can be stored in a file (precompiled) or directly in the DB.
Layouts must be a code file, but will change much less often.

Added config for development environment.

Removed the "default" route and just pointed it to show instead.

Added uids for sections, so I can style them differently.